Fix Operator Button Hover Effects
The operator buttons should be highlighted when the user is working
with it. However, the hover effect cancels this out. By creating
function handles to deal with the hover events, the event handlers can
be removed when the operator button is highlighted and re added when
the equal button is clicked (user is done with the current operator).