Fixed Hover Effect Again
The hover effect stopped working entirely after adding the higlhting
feature of operator buttons. This was because event handlers could not
be removed if it had an anonymous function but the function needed
the button parameter, not just the event parameter. By adding the this
keyword to the function handler instead of having an extra parameter,
the event handler can change the background color on hover and be
removed when the user highlights an operating button.